Remove add_annotation method
The add_annotation method is essentially the same as the add_layout method and
its somewhat confusing and unecessary to have both.

The method has never been in a release of Bokeh so doesn't need to be
deprecated.

We were already adding legends (which are annotations) with the add_layout
method, so I don't believe it will cause any confusion.

Further commits in this PR will enhance add_layout or annotations to make sure
the correct thing is done if someone attempts to do `add_layout(annotation,
'left')`.